requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
redis
pymongo
httpx
pytest
pytest-asyncio
PyYAML